+++ noodles-bgzf/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Added
 
-  * Add `libdeflate` feature to use [libdeflate] for decoding DEFLATE streams
-    (#51).
+  * Add `libdeflate` feature to use [libdeflate] for encoding and decoding
+    DEFLATE streams (#51).
 
 [#51]: https://github.com/zaeleus/noodles/issues/51
 [libdeflate]: https://github.com/ebiggers/libdeflate